requirements.txt
langchain-google-genai
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
llama-index-core
streamlit
plotly
pandas
